+++ packages/riverpod/lib/src/async_notifier.dart
@@ -60,7 +60,7 @@ abstract class AsyncNotifierBase<State> {
   /// {@endtemplate}
   Future<State> get future {
     _element.flush();
-    return _element._futureNotifier.value;
+    return _element.futureNotifier.value;
   }
 
   /// A function to update [state] from its previous value, while
@@ -139,7 +139,7 @@ ProviderElementProxy<AsyncValue<T>, Future<T>> _future<T>(
     that,
     (element) {
       return (element as AsyncNotifierProviderElement<AsyncNotifierBase<T>, T>)
-          ._futureNotifier;
+          .futureNotifier;
     },
   );
 }